pyproject.toml
[tool.poetry]
name = "diyims"
version = "0.0.0a24"
description = "A diy independent media services app that provides the modern version of the single person printing press."
authors = ["Douglas R Fix <douglasrfix@prairieobserver.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
repository = "https://github.com/douglasrfix/diyims"
packages = [{include = "diyims", from = "src"}]

[tool.poetry.scripts]
diyims = "diyims.diyims_cmd:app"

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
requests = ">=2.32.0"
aiosql = ">=9.0"
typer = ">=0.12.4"
python-dateutil = "^2.9.0.post0"
apscheduler = "^3.10.4"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    # "-m component",
    "-n 10",
    # "--dist load",
    "-rA",
    "-vv",
    "--capture=tee-sys",
    # "--cov=diyims",
    "--cov-append",
    "--cov-report=term-missing",
    "--basetemp=C:\\Users\\dougl\\diyims\\tmp",
]
markers = [
    "setup",
    "component",
    "mp",
]
[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.0.0"
pytest = ">= 8.0.0"
pytest-cov = ">=5.0.0"
flake8-pyproject = "^1.2.3"
ruff = "^0.7.0"
pytest-xdist = "^3.6.1"

[tool.flake8]
# B905 should be enabled when we drop support for 3.9
ignore = ['E203', 'E266', 'E501', 'E701', 'E704', 'W503', 'B905', 'B907']
# line length is intentionally set to 80 here because black uses Bugbear
# See https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#bugbear for more details
max-line-length = 80
max-complexity = 18
select = ['B', 'C', 'E', 'F', 'W', 'T4', 'B9']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
